refactor(#7): created controller and service package in test
